+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Cron class
 - Cron to clear image cache once every day to not litter thousands of images there
+- Warp Bubbles and Engeniering Complexes added to the killboard Plugin
 
 ## [v0.1-r20170307](https://github.com/ppfeufer/eve-online-wordpress-theme/releases/tag/v0.1-r20170307) - 2017-03-07
 [Full Changelog](https://github.com/ppfeufer/eve-online-wordpress-theme/compare/v0.1-r20170203...v0.1-r20170307)